Rewrite README in RDoc format.
`README` has been rewritten in the RDoc format. The wording in several sections
was massaged to be more understandable, but no major changes were made. Same
deal with code examples.

As an aside: the documentation and code samples in some sections are *really*
bad. For example, the "attachment_fu migrations" section is barely
comprehensible.